Stealth Mechanics
Escaping LOS and being quiet can make the NPCs lose you as a target and
go back to waypoint walking. They play a RAWR sound when finding you and
an error sound when losing you.
Also renamed them to MeleeAlien and duplicated and renamed the copy to
RangedAlien so I can start making shooting aliens!